requirement.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dominate==2.8.0
Flask==2.3.2
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
visitor==0.1.3
Werkzeug==2.3.5
WTForms==3.0.1
